ChatRoom Model added.
*The ChatRoom model was implemented to create 6 different chatrooms when
the server is launched.
*The socket logic was removed from the homepage. Sockets will only
launch a connection from a specific room page.
*The home page was modified to show the different chatrooms open for
the user. Users can click on the join button (or the room image) to be
redirected to the chatroom.
*A new route was created that validates if the chatroom exists.
*Finally, a util function to create the different chatrooms was created,
in the future this function will dissapear and an admin account must be
the only one authorized to create chatrooms dynamically.